refactor(ICache): move helpers into seperate file
