+++ src/main/scala/xiangshan/frontend/icache/IPrefetch.scala
@@ -70,7 +70,10 @@ class IPrefetchIO(implicit p: Parameters) extends IPrefetchBundle {
   val wayLookupWrite: DecoupledIO[WayLookupInfo] = DecoupledIO(new WayLookupInfo)
 }
 
-class IPrefetchPipe(implicit p: Parameters) extends IPrefetchModule with HasICacheECCHelper {
+class IPrefetchPipe(implicit p: Parameters) extends IPrefetchModule
+    with ICacheECCHelper
+    with ICacheAddrHelper {
+
   val io: IPrefetchIO = IO(new IPrefetchIO)
 
   private val (toITLB, fromITLB) = (io.itlb.map(_.req), io.itlb.map(_.resp))
